rename of example-widget and placed vue-widget inside the new renamed example-widget titled the whole thing vue-widget
