+++ apps/drupal/particle.info.yml
@@ -46,7 +46,6 @@ component-libraries:
       - ../../source/_patterns/02-molecules/jumbotron
       - ../../source/_patterns/02-molecules/pagination
       - ../../source/_patterns/02-molecules/vue-widget
-      - ../../source/_patterns/02-molecules/example-widget
   organisms:
     paths:
       - ../../source/_patterns/03-organisms
