+++ apps/grav/twig-namespaces.yaml
@@ -25,7 +25,6 @@ namespaces:
       - user/themes/particle/source/_patterns/02-molecules/jumbotron
       - user/themes/particle/source/_patterns/02-molecules/pagination
       - user/themes/particle/source/_patterns/02-molecules/vue-widget
-      - user/themes/particle/source/_patterns/02-molecules/example-widget
   organisms:
     paths:
       - user/themes/particle/source/_patterns/03-organisms/
