+++ apps/pl/pattern-lab/config/config.yml
@@ -75,7 +75,6 @@ plugins:
           - ../../../source/_patterns/02-molecules/jumbotron
           - ../../../source/_patterns/02-molecules/pagination
           - ../../../source/_patterns/02-molecules/vue-widget
-          - ../../../source/_patterns/02-molecules/example-widget
       organisms:
         paths:
           - ../../../source/_patterns/03-organisms
