+++ source/_patterns/02-molecules/vue-widget/__tests__/mock-data.json
