+++ source/_patterns/02-molecules/vue-widget/_example-widget.scss
