+++ source/_patterns/02-molecules/vue-widget/_vue-facet-table-vuex.twig
