--- source/_patterns/02-molecules/vue-widget/_vue-widget.twig
@@ -1,1 +1,0 @@
-<div id="vue-example-widget"></div>
