+++ source/_patterns/02-molecules/vue-widget/demo/example-widgets.twig
@@ -112,11 +112,11 @@ new Vue({
   <div class="row">
     <div class="col-sm-6">
       <h2>Vue</h2>
-      {% include '@molecules/example-widget/_vue-facet-table-vue.twig' %}
+      {% include '@molecules/vue-widget/_vue-facet-table-vue.twig' %}
     </div>
     <div class="col-sm-6">
       <h2>Vue + VueX</h2>
-      {% include '@molecules/example-widget/_vue-facet-table-vuex.twig' %}
+      {% include '@molecules/vue-widget/_vue-facet-table-vuex.twig' %}
     </div>
   </div>
