+++ source/_patterns/02-molecules/vue-widget/demo/vue-html-example.vue
