--- source/_patterns/02-molecules/vue-widget/demo/vue-widgets.twig
@@ -1,18 +1,0 @@
-{% extends '@atoms/grid/_grid-1-up.twig' %}
-  {% set container = true %}
-
-  {% block column_1 %}
-
-    <div class="card text-light bg-secondary mb-3">
-      <div class="card-body">
-        <h4 class="card-title">Vue Components</h4>
-        <p class="card-text">If my calculations are correct, the inner div should be mounted by the vueWidget app, which contains two components:</p>
-        <ul>
-          <li>a simple hello world message</li>
-          <li>a card component that re-uses _card.scss styles</li>
-        </ul>
-
-        {% include '@molecules/vue-widget/_vue-widget.twig' %}
-      </div>
-    </div>
-  {% endblock column_1 %}
