+++ source/_patterns/02-molecules/vue-widget/index.js
@@ -1,28 +1,41 @@
 /**
- * Example Vue widget
+ * Example Vue widgets
  *
- * An example of a testable, standalone, javascript-driven widget that has its own store
- * for state-based logic. That store can make whatever api calls you need it to.
- *
- * Note the use of this file (`vueWidget/index.js`) as the "implementation" of the actual
- * javascript application that resides in `vueWidget/src/index.js`. Double note that if you
- * are trying to see the Vue devtools in browser, you need to open the pattern outside
- * Pattern Lab's iframe.
+ * Mount all widgets via the root index.js file.
  */
 
-// Module dependencies
 import 'protons';
-import VueWidget from './src';
+import 'molecules/card';
 
-export const name = 'vueWidget';
+// Module template
+import './_vue-facet-table-vuex.twig';
+import './_vue-facet-table-vue.twig';
 
-export function enable() {
-  // Since the app is mounted as soon as possible in src/index.js, use enable()
-  // to pass in any data from outside apps (cough Drupal cough)
-  // @TODO: This always fires. Update to run only on visible.
-  VueWidget.$store.dispatch('vueWidget/getCardItems');
+// Module styles
+import './_example-widget.scss';
+
+import FacetTableVue from './src/vue-example'; // load only if dom found
+import FacetTableVueX from './src/vuex-example'; // load only if dom found
+import VueWidget from './src/vue-widget'; // load only if dom found
+
+// Render Vue elements as soon as possible
+if (document.getElementById('vue-facet-table-vue')) {
+  FacetTableVue('#vue-facet-table-vue');
+}
+if (document.getElementById('vue-facet-table-vuex')) {
+  FacetTableVueX('#vue-facet-table-vuex');
+}
+if (document.getElementById('vue-example-widget')) {
+  VueWidget('#vue-example-widget');
 }
 
+export const name = 'vue-widget';
+
 export function disable() {}
+
+export function enable() {
+  // Send data from settings or post-docready() work here
+  // FacetTableVue.$store.dispatch('exampleAction', settings.vueExampleData');
+}
 
 export default enable;
