+++ source/_patterns/02-molecules/vue-widget/src/vue-example/components/facet-table-facets.vue
