+++ source/_patterns/02-molecules/vue-widget/src/vue-example/facet-table.vue
