+++ source/_patterns/02-molecules/vue-widget/src/vue-widget/components/card.vue
