+++ source/_patterns/02-molecules/vue-widget/src/vuex-example/components/facet-table-facets.vue
