+++ source/_patterns/02-molecules/vue-widget/src/vuex-example/facet-table.vue
