+++ source/_patterns/02-molecules/vue-widget/src/vuex-example/index.js
